# fix: cache pending/ignore config lookups for faster perf/less memory