@fivetanley 2020-08-20
@rwjblue followed [your suggestions](https://github.com/ember-template-lint/ember-template-lint/pull/1466#discussion_r473141754) to move the new `ModuleStatusCache` to its own file (inside a new `-private` directory so people hopefully don't try to import it and use it directly) and assigning the instance in the `Linter` instance instead of in a weakmap in the `lib/get-config.js`'s module scope.

I re-ran the numbers on our app, here's the numbers (still an improvement! I didn't expect this to change but it seems good to verify that the refactor is working to improved speed).

I ran master and this PR branch 3 times each:

Before ([1c72839e544b2707a62cdada74c763ac274280cb](https://github.com/ember-template-lint/ember-template-lint/commit/1c72839e544b2707a62cdada74c763ac274280cb): 
```
$ /usr/bin/time ../ember-template-lint/bin/ember-template-lint.js . --quiet
       13.95 real        13.33 user         4.08 sys
$ /usr/bin/time ../ember-template-lint/bin/ember-template-lint.js . --quiet
       14.44 real        13.86 user         4.18 sys
$ /usr/bin/time ../ember-template-lint/bin/ember-template-lint.js . --quiet
       14.61 real        13.92 user         4.20 sys
```

After (this PR):

```
$ /usr/bin/time ../ember-template-lint/bin/ember-template-lint.js . --quiet
        9.80 real        11.90 user         0.94 sys
$ /usr/bin/time ../ember-template-lint/bin/ember-template-lint.js . --quiet
        9.12 real        11.51 user         0.86 sys
$ /usr/bin/time ../ember-template-lint/bin/ember-template-lint.js . --quiet
        9.09 real        11.43 user         0.85 sys
```